+++ tests/bottleneck.rs
@@ -178,7 +178,7 @@ fn bottleneck_no_overtaking_test() {
         assert_eq!(t, Time(exp_t), "Agent result: {:?}", agent_res);
     }
 
-    let weights = results.iteration_results.new_exp_weights.unwrap().clone();
+    let weights = results.iteration_results.new_exp_weights.clone();
     let weights = weights.road_network().unwrap();
     let edge_weight = &weights[(0, 0)];
     let TTF::Piecewise(ttf) = edge_weight else {
@@ -256,7 +256,7 @@ fn bottleneck_overtaking_test() {
         assert_eq!(t, Time(exp_t), "Agent result: {:?}", agent_res);
     }
 
-    let weights = results.iteration_results.new_exp_weights.unwrap().clone();
+    let weights = results.iteration_results.new_exp_weights.clone();
     let weights = weights.road_network().unwrap();
     let edge_weight = &weights[(0, 0)];
     let TTF::Piecewise(ttf) = edge_weight else {